Add files via upload
Site URL: http://www.stedwardsacademy.co.uk/ict-computing/
No Idea why it is coming up with the .co.uk domain ... we have been .org.o.uk for the 20 years that I have been there!!! We are a state funded school for 11-18 year olds.

We are currently moving to the delivery of Python for 13-18 year olds, initially with IDLE then hopefully PyCharm CE for those who will take coding as part of their exam. We also deliver HTML/CSS with a little javascript, where Dreamweaver is becoming a little long in the tooth ... not to mention the cumbersome UI.

I am currently trying to put together an intranet site to deliver some of this ... and Panic's Coda has just crashed again. 
Please let me know if there is any other information that I can provide.

Thanks.